requirements.txt
loadit
wandb
tqdm
wandb
hydra-core
einops
tqdm

# Below we use the stable versions for consistency.
torch==2.2.2
torchvision
# optax 0.2.2 uses scipy.linalg.tril, which is removed in scipy >1.13.0
scipy==1.12.0
optax==0.2.2
equinox==0.11.3
datasets==2.18.0
transformers==4.38.2
jaxamp

# we install jax and related packages last because otherwise pytorch will overwrite the CUDA installation with
# an older one and jax will complain that it cannot use the GPU
-f https://storage.googleapis.com/jax-releases/jax_cuda_releases.html "jax[cuda12_pip]" 